easy solution of stringConstruction
I think the easiest and more readable solution of stringConstruction challenge looks like this:
after splitting "b" string we map it and compare it to "a" string using the filter method. Then make flatten and save it in a variable. At last, just divide the length of this variable by the length of "a" string